so_long for macOS

+++ utils_bonus/animations/ft_delay_bonus.c
@@ -6,7 +6,7 @@
 /*   By: degabrie <degabrie@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/20 23:41:05 by degabrie          #+#    #+#             */
-/*   Updated: 2021/10/28 17:07:33 by degabrie         ###   ########.fr       */
+/*   Updated: 2021/12/06 21:13:16 by degabrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,6 @@ void	ft_delay(int ms)
 	pause = ms * (CLOCKS_PER_SEC / 1000);
 	then = clock();
 	now = then;
-	while ((now - then) < pause)
+	while ((now - then) < (unsigned long)pause)
 		now = clock();
 }
